構造の整理
