+++ .rspec
@@ -1,1 +1,3 @@
+--format documentation
+--color
 --require spec_helper
