+++ .travis.yml
@@ -1,0 +1,7 @@
+---
+sudo: false
+language: ruby
+cache: bundler
+rvm:
+  - 2.6.4
+before_install: gem install bundler -v 1.17.2
