--- lib/growi-image-converter-esa.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require 'growi-client'
-require 'open-uri'
-
-require 'growi-image-converter-esa/image_converter'
-require 'growi-image-converter-esa/markdown_image'
-require 'growi-image-converter-esa/attached_image_file'
-require 'growi-image-converter-esa/body'
-require 'growi-image-converter-esa/page'
