+++ lib/growi/image_converter/attached_image_file.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Growi
+  module ImageConverter
+    # Image Markdownとアタッチされた画像を対応させるクラス
+    class AttachedImageFile
+      def initialize(markdown_image, api_return_attached_file)
+        @markdown_image = markdown_image
+        @api_return_attached_file = api_return_attached_file
+      end
+      attr_accessor :markdown_image, :api_return_attached_file
+    end
+  end
+end
