+++ lib/growi/image_converter/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Growi
+  module ImageConverter
+    VERSION = '0.1.0'
+  end
+end
