+++ spec/growi/image_converter_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe Growi::ImageConverter do
+  it 'has a version number' do
+    expect(Growi::ImageConverter::VERSION).not_to be nil
+  end
+
+  it 'does something useful' do
+    expect(false).to eq(true)
+  end
+end
